scrapay-backend/requirements.txt
daphne==4.2.1
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.8
channels==4.3.2
channels-redis==4.3.0
python-dotenv==1.2.2
